style: rename folder to match naming convention
